requirements.txt
afip.py==1.1.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.4.0
bcrypt==4.3.0
black==25.1.0
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
ecdsa==0.19.1
exceptiongroup==1.3.0
frozenlist==1.8.0
geoip2==5.1.0
graphql-core==3.2.6
greenlet==3.2.2
h11==0.16.0
idna==3.10
inflect==7.5.0
iniconfig==2.1.0
maxminddb==2.8.2
more-itertools==10.7.0
multidict==6.7.0
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.4.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyodbc==5.2.0
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlacodegen==3.0.0
SQLAlchemy==2.0.41
starlette==0.46.2
strawberry-graphql==0.270.4
tomli==2.2.1
typeguard==4.4.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.5.0
uvicorn==0.34.2
watchfiles==1.1.0
websockets==15.0.1
winloop==0.1.91
yarl==1.22.0
